Fixed negative input out of bounds problem. Added some more things to mega-game class and changes to syntax here and tehre.

+++ EterTerminalDistribution/EterTerminalDistribution/ExplosionCard.h
@@ -30,5 +30,5 @@ public:
 	CardType GetCardType() const override;
 		
 	void RotateToRight(uint16_t);
-	void showExpl(size_t);
+	void showExpl(int16_t);
 };
